create osm subdir
